Addressing Debian bug #380551: Manual testing with emulated offline mode revealed several weaknesses in the URL selection process.
Do not display the $empty_warning after the user has picked a
continent or country.

Do not suggest "SPACE RETURN" when there is no previous pick anyway to
default to and "RETURN" alone will do just the same.

Tell the user to enter numbers (prevents the error in 380551
alltogether).

Change the help text if there is no urllist at all to say so.




git-svn-id: https://pause.perl.org:5460/svn/cpanpm/trunk@819 81f69c26-07c4-0310-b1c3-f486c8728884