Complain loadly if a Distroprefs pattern fail to compile
I happened to mention "$VERSION" in a comment in one of
my 01.DISABLED.yml lines.  The result was that all dists
where now silently disabled since the pattern now refuse
to compile and undef happen to match any distribution
name.

This patch makes it easy to spot this kind of error.