Major CPAN::FirstTime rewrite
First, all the MIRRORED.BY and "connect to internet" logic
was re-written to favor the most common case of someone
wanting to use the Internet.  The fast ("o conf init")
case is now parallel to "o conf init urllist" in terms
of prompting for the picklist.

Second, I shortened or reorganized a lot of the prompting
so that there is less text to read on screen for automatic
configuration.