Add support for HTTP::Lite
This commit adds support for HTTP::Lite as a pure-perl fallback
option if LWP is not installed.  HTTP::Lite (and MIME::Base64 and
Digest::MD5) have been added as prerequisites.

As a result of this change, HTTP::Lite will need to be added to the
Perl 5 core.  This will allow CPAN to bootstrap LWP (and the rest of
Bundle::CPAN) over http without reliance on external commands.  Jesse
Vincent has said this feature is a major goal for Perl 5.14.

The CPAN::HTTP::Lite module provides limited support for redirection and
authentication for HTTP::Lite, which does not natively support them,
and an API consistent with CPAN::LWP::UserAgent.

This patch changes auto-mirror selection to prefer 'http' over
'ftp' mirrors.