Warn and redo prompt if external program missing
When the user gives a path, complain if the path does not
point to an executable file, tell them how to disable it,
and repeat the prompt.

For 'make', be sure they see the warning about needing
make at least once.  That might be because it doesn't
exist in PATH.  If they haven't seen the warning, give
them another chance to configure make, even if they are
trying to disable it.  If they disable it after the warning,
let them.