Add function cpan_data_home
This function determines the "best" location for CPAN data files.

From candidate directories, we would like (in descending preference order):
   * the one that we loaded a MyConfig from
   * one that exists (even without MyConfig)
   * the first one on the list

This ensures that we always stay close to anything that the user had
previously.  This helps address the bug where upgrading File::HomeDir
causes files to be written FreeDesktop directories