Streamline configuration intro and auto_config logic
Now that we prompt about "using system defaults" in certain
circumstances (or even for missing data), the verbose intro
to configuration seems out of place.  The introduction
is replaced with a more action-oriented prompt.

As a side benefit, I've eliminated the "manual_config" variable
that had to be inverted to define auto_config.  The code is now
clearer.