Shortcut prereq_pm if no Makefile/Build exists
If configuration aborts with "exit 0", no Makefile and
no Build file will exist.  In such a case, we shouldn't
continue checking for prerequisites.

This avoids a late, confusing warning warning about being unable to read
META if a distribution has no META.  The warning was being triggered at
the end of queue processing, when color_cmd_tmps is called to reset
everything and calls prereq_pm.