fix prereq handling logic
This unbreaks configure_requires, which was broken somewhere in
refactoring.

The change makes satisfy_configure_requires and satisfy_requires
have similar return logic.

It changes how queuerunner deals with return values so that
regardless of the success/failure of the method, the item processed
gets removed from the queue.

If there are prereqs that need to be satisfied first, the item was
requeued anyway.