Fix ANSI output on Windows.
There's the fine option to create colorized output in CPAN. This uses
the Term::ANSIColor module, which is in core since perl 5.6.0.

However on Windows, your typical 'cmd.exe' session can not display
colors all that well. For that, you must first load the
Win32::Console::ANSI module, which is not in core.

I added a check if this module is available (only if run on Windows)
and turn colors off if it is not. Also I extended the option text you
see when doing 'o conf init' to include a reference to the module.

This fixes RT48909 and its duplicate RT54931.