Use JSON as fallback for YAML format for reports.
The 'YAML' module has problems decoding the reports from
cpantesters - the 'reports' function in CPAN only works if you have
YAML::XS or YAML::Syck installed.

There is another issue, which is that all these modules - including
YAML - are not core modules and cpan does not warn you if you have
not installed them.

As a workaround to the YAML parsing problem, we now only download
YAML if you have YAML::XS or YAML::Sick, otherwise we download the
JSON version of the report. It parses this with JSON::PP - which is
in core since 5.14. Of course I understand not all of us are lucky
enough to have a relatively new perl such as that, so if you don't
have any of these modules 'reports' will warn you about it and stop.