Compare 'undef' versions as '0'
version.pm treats the string "undef" as 0.  And "use Foo 0" will
allow undefined $VERSION.  Therefore, this commit changes any
literal "undef" versions passed to vcmp to 0 before comparision.

This fixes a bug in CPAN::Module::uptodate for modules that have
version "0" in the package list and do not define a $VERSION
in the .pm file.  Previously, it would always return false.  This
makes it return true.