Fix Makefile/Build file test in prereq_pm.
e55b963a6b8c8 added a check to look for 'Makefile' or 'Build' and stopped
looking for prerequisites if those files don't exist.  That means no
prerequisite processing is done for any make tool or platform that doesn't
generate those specific filenames.  So add at least a couple of the things
that are missing.